backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
supabase==2.3.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.1
